[Refactor] 컴포넌트 분리
- 차트, 필터, 툴팁 컴포넌트 분리
